Update alignment on account names/icons; add a missing Q_OBJECT macro to witness plot tracker widget.
